54/requirements.txt
aiogram>=3.0.0
aiosqlite
python-dotenv
pyyaml
